requirements.txt
pytube
pydub
requests
openai
tiktoken
chromadb
langchain
python-dotenv
pyAudioAnalysis
eyed3
scikit-learn